Adds a new hook to tweak command description at runtime
For one feature I'm currently implementing I need to be able to tweak a command description based on some runtime information.
My main use case is to customize the description (basically rendering it as a template but with some context values that we can't provide generically).
So I ended up with this solution.

The idea is to allow one to register a hook that will be called during command help rendering to compute the command description.